MAP 257 - Update message builder with optional fields of motor, bike and crosswalk layers (#97)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description
In this PR
- Updated message builder with optional fields of motor vehicle, bike,
and crosswalk layers
- Updated IntersectionInput.java file to support new optional fields
- Fixed a bug in JNI wrapper
- Created new unit test cases and updated old unit test cases to test
the new code

<!--- Describe your changes in detail -->

## Related GitHub Issue

<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key
[MAP-257](https://usdot-carma.atlassian.net/browse/MAP-257)
<!-- e.g. CAR-123 -->

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Defect fix (non-breaking change that fixes an issue)
- [x] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.


[MAP-257]:
https://usdot-carma.atlassian.net/browse/MAP-257?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

+++ fedgov-cv-rgaencoder/src/main/java/gov/usdot/cv/rgaencoder/GeneralPeriod.java
@@ -17,7 +17,7 @@
 package gov.usdot.cv.rgaencoder;
 
 public class GeneralPeriod {
-    private short generalPeriodValue;
+    private int generalPeriodValue;
 
     // Constants for choice field
     public static final int DAY = 0;
@@ -26,7 +26,7 @@ public class GeneralPeriod {
     public GeneralPeriod() {
     }
 
-    public GeneralPeriod(short generalPeriodValue) {
+    public GeneralPeriod(int generalPeriodValue) {
         this.generalPeriodValue = generalPeriodValue;
     }
 
@@ -34,7 +34,7 @@ public class GeneralPeriod {
         return generalPeriodValue;
     }
 
-    public void setGeneralPeriodValue(short generalPeriodValue) {
+    public void setGeneralPeriodValue(int generalPeriodValue) {
         this.generalPeriodValue = generalPeriodValue;
     }
